Made it so enemy detects if it hits player
